+++ packages/core/navigation-next/docs/20-composing-your-navigation.js
@@ -68,7 +68,7 @@ ${(
 
 ${<H>Composing the product navigation</H>}
 
-Let's move on to the next part of our navigation - the product area. This area will always contain the product layer. Depending on the part of the application we're in there may also be a container layer as well. The \`productContainerNavigation\` prop is nullable, so we'll remove it for now. We'll also create a component to render something in the product area.
+Let's move on to the next part of our navigation - the product area. This area will always contain the product layer. Depending on the part of the application we're in there may also be a container layer as well. The \`containerNavigation\` prop is nullable, so we'll remove it for now. We'll also create a component to render something in the product area.
 
 We can render basically anything we want in this section. All of the typical primitive UI components are provided by \`@atlaskit/navigation-next\` so let's start putting a few of these together.
 
@@ -106,7 +106,7 @@ const MyProductNavigation = () => (
 
 **Note:** We don't automatically apply padding or add wrapping elements around your product navigation or within \`Section\`s because we don't want to break your layout. Instead, we provide the \`Section\`'s children with styles through a render function.
 
-Read the Component docs for more information about the \`Section\`, along with every other UI primitive exported by this package.
+Read the [component props docs](/packages/core/navigation-next/docs/ui-components) for more information about the \`Section\`, along with every other UI primitive exported by this package.
 
 Putting it all together we can now render a somewhat complete navigation!
 
@@ -132,23 +132,23 @@ ${(
 
 ${<H>Managing the UI state</H>}
 
-If you try you'll notice that you can expand, collapse, and resize your navigation right now, and this state will be controlled by the component. What if we want to manage or respond to this state somewhere in our application? Introducing the \`UIStateSubscriber\`.
+If you try you'll notice that you can expand, collapse, and resize your navigation right now, and this state will be controlled by the component. What if we want to manage or respond to this state somewhere in our application? Introducing the \`UIControllerSubscriber\`.
 
 ${code`
-import { UIStateSubscriber } from '@atlaskit/navigation-next';
+import { UIControllerSubscriber } from '@atlaskit/navigation-next';
 
 const ExpandToggleButton = () => (
-  <UIStateSubscriber>
+  <UIControllerSubscriber>
     {navigationUIController => (
       <button onClick={navigationUIController.toggleCollapse}>
-        {navigationUIController.state.isCollapsed ? 'Expand' : 'Collapse'} product navigation
+        {navigationUIController.state.isCollapsed ? 'Expand' : 'Collapse'} content navigation
       </button>
     )}
-  </UIStateSubscriber>
+  </UIControllerSubscriber>
 );
 `}
 
-This component provides its children with access to the state of the navigation, along with methods which can be called to modify the state. It will re-render whenever the state changes. This container manages behaviours such as collapsing/expanding, peeking, peek hinting, and resizing. Read the component's documentation for a complete list of methods and properties.
+This component provides its children with access to the state of the navigation, along with methods which can be called to modify the state. It will re-render whenever the state changes. This container manages behaviours such as collapsing, expanding, and resizing. Read [the component's documentation](/packages/core/navigation-next/docs/state-controllers#ui-controller) for a complete list of methods and properties.
 
 ${(
     <IframeExample