+++ packages/core/navigation-next/src/components/ContentNavigation/primitives.js
@@ -1,6 +1,6 @@
 // @flow
 
-import React, { type Node } from 'react';
+import React, { Fragment, type Node } from 'react';
 import { keyframes } from 'emotion';
 import { ThemeProvider } from 'emotion-theming';
 import { Transition } from 'react-transition-group';
@@ -40,6 +40,8 @@ const ScrollProvider = (props: any) => (
   <div
     css={{
       boxSizing: 'border-box',
+      display: 'flex',
+      flexDirection: 'column',
       height: '100%',
       overflowX: 'hidden',
       overflowY: 'auto',
@@ -80,7 +82,7 @@ export const ProductNavigationTheme = ({ children }: BaseNavigationTheme) => (
   <ThemeProvider
     theme={oldTheme => ({ mode: light, ...oldTheme, context: 'product' })}
   >
-    {children}
+    <Fragment>{children}</Fragment>
   </ThemeProvider>
 );
 
@@ -152,7 +154,7 @@ type ContainerNavigationProps = {
 
 export const ContainerNavigationTheme = ({ children }: BaseNavigationTheme) => (
   <ThemeProvider theme={{ mode: light, context: 'container' }}>
-    {children}
+    <Fragment>{children}</Fragment>
   </ThemeProvider>
 );