+++ packages/core/navigation-next/src/components/LayoutManagerWithViewController/__tests__/unit/LayoutManagerWithViewController.js
@@ -5,13 +5,9 @@ import { mount } from 'enzyme';
 import { HashRouter } from 'react-router-dom';
 
 import LayoutManagerWithViewController from '../../LayoutManagerWithViewController';
-
-import {
-  DefaultGlobalNavigation,
-  ProjectSwitcher,
-} from '../../../../../examples/shared/components';
-
 import { NavigationProvider } from '../../../../index';
+
+const GlobalNavigationComponent = () => null;
 
 describe('LayoutManagerWithViewController', () => {
   let wrapper;
@@ -21,6 +17,7 @@ describe('LayoutManagerWithViewController', () => {
   let onCollapseEnd;
   let onExpandStart;
   let onExpandEnd;
+  let getRefs;
 
   beforeEach(() => {
     originalLocalStorage = global.localStorage;
@@ -40,6 +37,7 @@ describe('LayoutManagerWithViewController', () => {
     onCollapseEnd = jest.fn();
     onExpandStart = jest.fn();
     onExpandEnd = jest.fn();
+    getRefs = jest.fn();
 
     wrapper = mount(
       <HashRouter>
@@ -48,13 +46,13 @@ describe('LayoutManagerWithViewController', () => {
           isDebugEnabled={false}
         >
           <LayoutManagerWithViewController
-            customComponents={{ ProjectSwitcher }}
-            globalNavigation={DefaultGlobalNavigation}
+            globalNavigation={GlobalNavigationComponent}
             firstSkeletonToRender={'product'}
             onCollapseStart={onCollapseStart}
             onCollapseEnd={onCollapseEnd}
             onExpandStart={onExpandStart}
             onExpandEnd={onExpandEnd}
+            getRefs={getRefs}
           >
             <p>
               Children requires to have `NavigationProvider` as a parent Because
@@ -76,10 +74,11 @@ describe('LayoutManagerWithViewController', () => {
     onCollapseEnd.mockReset();
     onExpandStart.mockReset();
     onExpandEnd.mockReset();
+    getRefs.mockReset();
   });
 
   it('should render global navigation based on using `globalNavigation` as a reference', () => {
-    expect(wrapper.find(DefaultGlobalNavigation).length).toBe(1);
+    expect(wrapper.find(GlobalNavigationComponent).length).toBe(1);
   });
 
   describe('LayerInitialised', () => {
@@ -118,8 +117,7 @@ describe('LayoutManagerWithViewController', () => {
             isDebugEnabled={false}
           >
             <LayoutManagerWithViewController
-              customComponents={{ ProjectSwitcher }}
-              globalNavigation={DefaultGlobalNavigation}
+              globalNavigation={GlobalNavigationComponent}
               firstSkeletonToRender={'container'}
             >
               <p>
@@ -145,8 +143,8 @@ describe('LayoutManagerWithViewController', () => {
     });
   });
 
-  describe('Passing Props to LayerManager', () => {
-    it('should pass Collapse Listeners props', () => {
+  describe('Passing props to LayoutManager', () => {
+    it('should pass expand/collapse listeners and getRefs', () => {
       const layoutManager = wrapper.find('LayoutManager');
 
       onCollapseStart(200);
@@ -158,6 +156,7 @@ describe('LayoutManagerWithViewController', () => {
       expect(layoutManager.props().onCollapseEnd).toBeCalledWith(0);
       expect(layoutManager.props().onExpandStart).toBeCalledWith(0);
       expect(layoutManager.props().onExpandEnd).toBeCalledWith(200);
+      expect(layoutManager.props().getRefs).toHaveBeenCalled();
     });
   });
 });