+++ packages/core/radio/src/index.js
@@ -1,5 +1,4 @@
 // @flow
 export { default as RadioGroup } from './RadioGroup';
 export { default as RadioIcon } from './RadioIcon';
-export { default as RadioInput } from './RadioInput';
 export { default as Radio } from './Radio';