+++ packages/editor/editor-wikimarkup-transformer/src/__tests__/unit/parser/file-link.ts
@@ -1,0 +1,14 @@
+import WikiMarkupTransformer from '../../../index';
+
+describe('JIRA wiki markup - File Link', () => {
+  const testCases: Array<[string, string]> = [
+    ['should find file link in the text', '[^file.pdf]'],
+  ];
+
+  for (const [testCaseDescription, markup] of testCases) {
+    it(testCaseDescription, () => {
+      const transformer = new WikiMarkupTransformer();
+      expect(transformer.parse(markup)).toMatchSnapshot();
+    });
+  }
+});