+++ website/src/pages/ChangeLogExplorer.js
@@ -1,12 +1,12 @@
 // @flow
 import React, { Component, type Node } from 'react';
 import styled from 'styled-components';
-import { Link } from 'react-router-dom';
+import { Link } from '../components/WrappedLink';
 
 import BackIcon from '@atlaskit/icon/glyph/arrow-left';
 import TextField from '@atlaskit/field-text';
 import Button from '@atlaskit/button';
-import Loadable from 'react-loadable';
+import Loadable from '../components/WrappedLoader';
 
 import Changelog, { NoMatch } from '../components/ChangeLog';
 import Page from '../components/Page';