+++ packages/editor/editor-jira-transformer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/editor-jira-transformer
 
+## 5.0.4
+- [patch] Update TS to 3.0 [f68d367](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f68d367)
+- [none] Updated dependencies [f68d367](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f68d367)
+  - @atlaskit/editor-json-transformer@4.0.10
+  - @atlaskit/editor-common@15.0.7
+  - @atlaskit/editor-test-helpers@6.0.8
+  - @atlaskit/editor-core@79.0.12
+
 ## 5.0.3
 - [patch] Updated dependencies [7545979](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/7545979)
   - @atlaskit/editor-common@15.0.0