+++ packages/editor/editor-test-helpers/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/editor-test-helpers
 
+## 6.0.8
+- [patch] Update TS to 3.0 [f68d367](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f68d367)
+- [none] Updated dependencies [f68d367](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f68d367)
+  - @atlaskit/media-test-helpers@17.0.2
+  - @atlaskit/media-core@23.1.1
+  - @atlaskit/editor-common@15.0.7
+  - @atlaskit/editor-core@79.0.12
+
 ## 6.0.7
 - [patch] ED-4680 add smart card plugin, enable for inline smart cards [b9529e6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b9529e6)